updated readme, moved passport instance creation to utils file
