+++ examples/passport/wallets-with-nextjs/src/app/connect-with-eip1193/page.tsx
@@ -2,7 +2,7 @@
  
 import { useState } from 'react';
 import { ProviderEvent } from '@imtbl/sdk/passport';
-import { passportInstance } from '../page';
+import { passportInstance } from '../utils';
 
 export default function ConnectWithEtherJS() {
